signal_aggregator/requirements.txt
aiokafka>=0.10.0
kafka-python>=2.0.2
asyncpg>=0.29.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
pydantic>=2.0.0
PyYAML>=6.0
prometheus-client>=0.19.0